requirements.txt
python-telegram-bot==20.0
requests==2.28.1
python-dotenv==1.0.0
Flask==3.0.0